src/main/java/com/luv2code/springmvc/service/StudentAndGradeService.java
package com.luv2code.springmvc.service;

import com.luv2code.springmvc.models.CollegeStudent;
import com.luv2code.springmvc.repository.StudentRepository;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.Optional;

@Service
@Transactional
@AllArgsConstructor
public class StudentAndGradeService {

    private final StudentRepository studentRepository;

    public void createStudent(String firstName, String lastName, String email) {
        CollegeStudent student = new CollegeStudent(firstName, lastName, email);
        student.setId(0);
        studentRepository.save(student);
    }

    public Boolean checkIfStudentIsNull(Integer id) {
        Optional<CollegeStudent> student = studentRepository.findById(id);
        return student.isPresent();
    }

    public void deleteStudent(Integer id) {
        studentRepository.deleteById(id);
    }

    public Iterable<CollegeStudent> getGradeBook() {
        Iterable<CollegeStudent> collegeStudents = studentRepository.findAll();
        return collegeStudents;
    }

}
